backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==1.10.8
python-dotenv==1.0.0
weaviate-client==4.10.4
beautifulsoup4==4.12.2
requests==2.31.0
sentence-transformers==2.2.2
scikit-learn==1.3.1
langchain==0.92.1
numpy==1.26.0
